+++ practice-tool/src/config.rs
@@ -40,8 +40,6 @@ pub(crate) struct Settings {
     #[serde(default)]
     pub(crate) show_console: bool,
     #[serde(default)]
-    pub(crate) disable_install_prompt: bool,
-    #[serde(default)]
     pub(crate) disable_update_prompt: bool,
     #[serde(default = "Indicator::default_set")]
     pub(crate) indicators: Vec<Indicator>,
@@ -295,7 +293,6 @@ impl Default for Config {
                 dxgi_debug: false,
                 show_console: false,
                 indicators: Indicator::default_set(),
-                disable_install_prompt: false,
                 disable_update_prompt: false,
             },
             commands: Vec::new(),
