+++ practice-tool/src/practice_tool.rs
@@ -14,7 +14,7 @@ use practice_tool_core::widgets::{scaling_factor, Widget, BUTTON_HEIGHT, BUTTON_
 use tracing_subscriber::prelude::*;
 
 use crate::config::{Config, Indicator, Settings};
-use crate::install::{Install, Update};
+use crate::update::Update;
 use crate::util;
 
 const MAJOR: usize = pkg_version_major!();
@@ -49,7 +49,6 @@ pub(crate) struct PracticeTool {
     fonts: Option<FontIDs>,
     config_err: Option<String>,
     update_available: Update,
-    install_necessary: Install,
 
     position_bufs: [String; 4],
     igt_buf: String,
@@ -161,12 +160,6 @@ impl PracticeTool {
         let update_available =
             if config.settings.disable_update_prompt { Update::UpToDate } else { Update::check() };
 
-        let install_necessary = if config.settings.disable_install_prompt {
-            Install::Unnecessary
-        } else {
-            Install::check()
-        };
-
         let pointers = Pointers::new();
         let version_label = {
             let (maj, min, patch) = (*VERSION).into();
@@ -191,7 +184,6 @@ impl PracticeTool {
             position_bufs: Default::default(),
             igt_buf: Default::default(),
             update_available,
-            install_necessary,
         }
     }
 
@@ -274,30 +266,6 @@ impl PracticeTool {
                     ui.open_popup("##help_window");
                 }
 
-                match &self.install_necessary {
-                    Install::Unnecessary => {},
-                    Install::Necessary { .. } => {
-                        ui.same_line();
-
-                        let green = [0.1, 0.7, 0.1, 1.0];
-                        let _token = ui.push_style_color(StyleColor::Button, green);
-
-                        if ui.small_button("Install") {
-                            ui.open_popup("##install");
-                        }
-                    },
-                    Install::Error(_) => {
-                        ui.same_line();
-
-                        let red = [1.0, 0.0, 0.0, 1.0];
-                        let _token = ui.push_style_color(StyleColor::Button, red);
-
-                        if ui.small_button("Install") {
-                            ui.open_popup("##install");
-                        }
-                    },
-                }
-
                 match &self.update_available {
                     Update::UpToDate => {},
                     Update::Available { .. } => {
@@ -362,40 +330,6 @@ impl PracticeTool {
                             open::that("https://patreon.com/johndisandonato").ok();
                         }
                         ui.same_line();
-                        if ui.button("Close") {
-                            ui.close_current_popup();
-                            self.pointers.cursor_show.set(false);
-                        }
-                    });
-
-                ui.modal_popup_config("##install")
-                    .resizable(false)
-                    .movable(false)
-                    .title_bar(false)
-                    .build(|| {
-                        self.pointers.cursor_show.set(true);
-
-                        match &self.install_necessary {
-                            Install::Unnecessary => ui.text("Install successful."),
-                            Install::Necessary { .. } => {
-                                ui.text(
-                                    "Do you want to install the tool?\nThis will copy the tool's \
-                                     DLL file next to a file \nnamed `dinput8.dll` right next to \
-                                     `eldenring.exe`.\nIf you want to uninstall the tool, all you \
-                                     have to\ndo is remove that file.",
-                                );
-                                if ui.button("Install") {
-                                    self.install_necessary.install();
-                                }
-                                ui.same_line();
-                            },
-                            Install::Error(e) => {
-                                ui.text("Error while trying to determine installation status.");
-                                ui.separator();
-                                ui.text(e);
-                            },
-                        }
-
                         if ui.button("Close") {
                             ui.close_current_popup();
                             self.pointers.cursor_show.set(false);
